refactor: extract ProjectDialog and TimelineDialog from AdminTabs
Form submit logic and form JSX are co-located in their own components.
AdminTabs is now a coordinator: state, drag-drop, archive, delete, and
wiring only. Also collapse tab buttons to a .map() to remove duplication.
